Add google-maps-react-hooks
