+++ src/page/home/index.jsx
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <h1>Here is Home</h1>;
+};
+
+export default Home;
